refactor: 调整 workspace 配置顺序
