+++ pnpm-workspace.yaml
@@ -1,3 +1,5 @@
+shellEmulator: true
+
 catalogs:
   frontend:
     '@tanstack/query-persist-client-core': ^5.91.4
@@ -56,7 +58,3 @@ catalogs:
     vite: ^4.5.14
   typing:
     '@types/lodash-es': ^4.17.12
-
-cleanupUnusedCatalogs: true
-
-shellEmulator: true
